requirements.txt
streamlit
langchain
faiss-cpu
python-docx
openai
tiktoken
langchain-community